azure-function/requirements.txt
azure-functions
aiohttp>=3.8.0
asyncio
